+++ packages/react/src/DevMode.tsx
@@ -1,3 +1,16 @@
+import type { ComponentProps, ComponentType } from 'react'
+
+export type PropsWithDevMode<
+  TProps extends ComponentProps<ComponentType>,
+  TDevModeProps extends Record<string, unknown>,
+> = TProps & {
+  /**
+   * @deprecated Use official react devtools instead
+   * @see https://react.dev/learn/react-developer-tools
+   */
+  devMode?: TDevModeProps
+}
+
 interface DevModeProps {
   /**
    * @deprecated Use official react devtools instead
