--- packages/react/src/utility-types/index.ts
@@ -1,2 +1,0 @@
-export type { ConstructorType } from './ConstructorType'
-export type { PropsWithDevMode } from './PropsWithDevMode'
