+++ packages/utility-types/src/index.ts
@@ -1,2 +1,3 @@
 export type { OmitKeyof } from './OmitKeyof'
 export type { RequiredKeyof } from './RequiredKeyof'
+export type { ConstructorType } from './ConstructorType'
